# enterprise/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Boundless Enterprise Multipass API
  description: |
    Enterprise-grade unified access and control layer for the Boundless blockchain platform.

    ## Features
    - Post-Quantum Cryptography (Dilithium5 + Kyber1024)
    - KYC/AML Identity Management
    - Multi-Asset Wallet System
    - Internal Asset Trading Market
    - Hardware Pass NFC Integration
    - SSO & Session Management
    - Enterprise Application Registry
    - Events & Reporting

    ## Security
    - All endpoints (except `/api/auth/login` and `/api/auth/register`) require authentication
    - JWT-based session tokens
    - Rate limiting on sensitive endpoints
    - Encrypted private key storage with master encryption key

    ## Base URL
    `http://localhost:8080` (default development)

  version: 1.0.0
  contact:
    name: Boundless Platform
    url: https://boundless.example.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.boundless.example.com
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from `/api/auth/login`
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required:
        - error
    Identity:
      type: object
      properties:
        identity_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
        kyc_status:
          type: string
          enum: [Pending, Verified, Rejected, Expired]
        created_at:
          type: string
          format: date-time
...
    Session:
      type: object
      properties:
        session_id:
          type: string
          format: uuid
        identity_id:
          type: string
          format: uuid
        token:
          type: string
          description: JWT authentication token
        expires_at:
          type: string
          format: date-time
paths:
...
  /api/auth/register:
    post:
      tags: [Authentication]
      summary: Register new user
      description: Create new identity and authentication credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, full_name]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                full_name:
                  type: string
      responses:
        '200':
          description: Registration successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  identity_id:
                    type: string
                    format: uuid
                  session:
                    $ref: '#/components/schemas/Session'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: Login user
      description: Authenticate with email and password, returns JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many login attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/logout:
    post:
      tags: [Authentication]
      summary: Logout user
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session_id]
              properties:
                session_id:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Logout successful
        '401':
          description: Unauthorized

  # =============================================================================
  # IDENTITY
  # =============================================================================
...
  /api/identity/{id}:
    get:
      tags: [Identity]
      summary: Get identity by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Identity profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Identity'
...
  /api/hardware/register:
    post:
      tags: [Hardware]
      summary: Register NFC hardware device
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identity_id, device_name, hardware_id]
              properties:
                identity_id:
                  type: string
                  format: uuid
                device_name:
                  type: string
                hardware_id:
                  type: string
                  description: Unique hardware identifier
      responses:
        '200':
          description: Device registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  device_id:
                    type: string
                    format: uuid
  /api/hardware/{device_id}/authenticate:
    post:
      tags: [Hardware]
      summary: Authenticate with hardware device
      description: Challenge-response authentication with NFC device
      security:
        - BearerAuth: []
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challenge_response]
              properties:
                challenge_response:
                  type: string
                  description: Signed challenge from hardware device
      responses:
        '200':
          description: Authentication successful
security:
  - BearerAuth: []